refactor: split gorialize into multiple files
